Customization

+++ server/src/utility/deleteFiles.js
@@ -1,7 +1,7 @@
 const fs = require("fs");
 const path = require("path");
 
-function deleteFiles(subFolder) {
+function deleteFiles(subFolder, except) {
   const directory = path.join(`${__dirname}/../../public/${subFolder}`);
   fs.readdir(directory, (error, files) => {
     if (error) {
@@ -10,13 +10,15 @@ function deleteFiles(subFolder) {
       });
     } else {
       for (const file of files) {
-        fs.unlink(path.join(directory, file), (error) => {
-          if (error) {
-            res.status(500).json({
-              error: error.message,
-            });
-          }
-        });
+        if(file!=except){
+          fs.unlink(path.join(directory, file), (error) => {
+            if (error) {
+              res.status(500).json({
+                error: error.message,
+              });
+            }
+          });
+        }
       }
     }
   });
